Cambios Pantalla Detalles

+++ app/src/main/java/com/example/edu/a0817moacn01c_3/DetalleFragment.java
@@ -36,7 +36,7 @@ public class DetalleFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         Bundle unBundle = getArguments();
-        pelicula = new Peliculas(unBundle.getString("nombre"),unBundle.getInt("imagen"),unBundle.getInt("imagenPortada"),unBundle.getString("genero"),unBundle.getString("descripcion"),unBundle.getInt("puntuacion"),unBundle.getString("aptoTodoPublico"));
+        pelicula = new Peliculas(unBundle.getString("nombre"),unBundle.getInt("imagen"),unBundle.getInt("imagenPortada"),unBundle.getString("genero"),unBundle.getString("descripcion"),unBundle.getDouble("puntuacion"),unBundle.getString("aptoTodoPublico"));
         // Inflate the layout for this fragment
         View view = inflater.inflate(R.layout.fragment_detalle, container, false);
         this.imagen = view.findViewById(R.id.imageView_ImagenPelicula);
